Day-3/pyproject.toml
[project]
name = "day-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.29",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.9",
    "openai>=1.99.5",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
]
